getting weather information for Chicago + test in main
