+++ src/main/java/com/group2/project/Main.java
@@ -1,5 +1,7 @@
 package com.group2.project;
 
+import com.group2.project.Weather.Weather;
+import com.group2.project.Weather.WeatherData;
 import javafx.application.Application;
 import org.apache.catalina.Context;
 import org.apache.tomcat.util.scan.StandardJarScanner;
@@ -18,6 +20,14 @@ public class Main
     public static void main(String[] args) {
         //SpringApplication.run(Main.class, args);
         ProjectApplication.main(args);
+
+        Weather weather = new Weather();
+        WeatherData weatherData = weather.getWeatherData();
+        if (weatherData != null) {
+            System.out.println(weatherData);
+        } else {
+            System.out.println("Weather data could not be retrieved.");
+        }
     }
